updated realDistance function to show int

+++ js/index.js
@@ -111,5 +111,5 @@ function keyReleased() {
 }
 
 function realDistance() {
-  return distanceWalked / DISTANCE_COEFFICIENT;
+  return int(distanceWalked / DISTANCE_COEFFICIENT);
 }
